refactor: NavigationMobile and NavigationDesktop components
